requirements/base.txt
fastapi==0.73.0
pydantic==1.9.0
uvicorn==0.17.4
python-dotenv==0.19.2
sentry-sdk==1.5.6
zeep==4.1.0
requests==2.27.1
PyJWT==2.3.0
pycryptodome==3.14.1
redis==4.2.2
datadog==0.44.0
